- Move resumable file upload outside CAS filters purview  - Add authz to resumable file upload  - Remove unused file upload actions from TaskController
