+++ grails-app/controllers/au/org/ala/volunteer/AjaxController.groovy
@@ -10,6 +10,8 @@ import com.google.common.collect.Sets
 import com.google.gson.Gson
 import grails.converters.JSON
 import grails.converters.XML
+import groovy.util.logging.Slf4j
+import org.springframework.web.multipart.MultipartHttpServletRequest
 
 import java.sql.Timestamp
 import java.text.SimpleDateFormat
@@ -22,7 +24,16 @@ import grails.plugins.csv.CSVWriterColumnsBuilder
 import java.util.concurrent.TimeUnit
 
 import static grails.async.Promises.*
+import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST
+import static javax.servlet.http.HttpServletResponse.SC_CREATED
+import static javax.servlet.http.HttpServletResponse.SC_FORBIDDEN
+import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR
+import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND
+import static javax.servlet.http.HttpServletResponse.SC_NO_CONTENT
+import static javax.servlet.http.HttpServletResponse.SC_PRECONDITION_FAILED
+import static javax.servlet.http.HttpServletResponse.SC_UNAUTHORIZED
 
+@Slf4j
 class AjaxController {
 
     def taskService
@@ -609,8 +620,6 @@ class AjaxController {
         respond results
     }
 
-
-
     private static def getNamedFieldValues(List fields, String name) {
         fields?.findAll { it.name == name && it.value }?.sort { it.recordIdx }?.collect { it.value }?.join(', ')
     }
@@ -627,5 +636,66 @@ class AjaxController {
             result = format.parse(timestamp).toTimestamp()
         }
         return result
+    }
+
+
+    def resumableUploadFile(ResumableUploadCommand cmd) {
+
+        if (cmd.hasErrors()) {
+            log.error("Resumable params are not valid {}", cmd)
+            return render(status: SC_BAD_REQUEST, text: "Params aren't valid")
+        }
+
+        def allowedMimeTypes = ['image/jpeg', 'image/gif', 'image/png', 'text/plain']
+        if (!allowedMimeTypes.contains(cmd.type)) {
+            log.error("Resumable file content-type is not valid {}", cmd)
+            return render(status: SC_BAD_REQUEST, text: "The image file must be one of: ${allowedMimeTypes}")
+        }
+
+        def project = Project.get(cmd.projectId)
+        if (!project) {
+            return render(status: SC_NOT_FOUND, text: "Project doesn't exist")
+        }
+
+        if (!projectService.isAdmin(project, request.userPrincipal, request.&isUserInRole)) {
+            return render(status: request.userPrincipal ? SC_FORBIDDEN : SC_UNAUTHORIZED, text: 'Access denied')
+        }
+
+        log.debug("Uploading {}:{} identifier {} size {} checksum {}", cmd.filename, cmd.resumableChunkNumber, cmd.identifier, cmd.resumableCurrentChunkSize, cmd.checksum)
+
+        if (request.method == 'POST') {
+            try {
+                def resumableStream = getResumableStream()
+                if (resumableStream != null) {
+                    def chunkCheck = cmd.uploadAndCheckChunk(resumableStream)
+                    if (chunkCheck) {
+                        render status: SC_CREATED, text: ''
+                    } else {
+                        render status: SC_PRECONDITION_FAILED, text: 'Chunk checksum does not match'
+                    }
+                } else {
+                    render status: SC_BAD_REQUEST, text: 'No file part found'
+                }
+            } catch (e) {
+                log.error("Couldn't save uploaded chunk {}", cmd, e)
+                render status: SC_INTERNAL_SERVER_ERROR, text: "Couldn't save uploaded chunk ${cmd.filename}:${cmd.resumableChunkNumber}"
+            }
+        } else if (request.method == 'GET') {
+            if (cmd.isChunkComplete()) {
+                render(status: SC_NO_CONTENT, text: '')
+            } else {
+                render(status: SC_NOT_FOUND, text: '')
+            }
+        }
+    }
+
+    private getResumableStream() {
+        if (request instanceof MultipartHttpServletRequest) {
+            // theoretically there's only one chunk being sent in this request, so just find the first file part
+            def mfm = request.getMultiFileMap()
+            mfm.values().collectMany { it }.find()?.inputStream
+        } else {
+            request.inputStream
+        }
     }
 }
